Moved legal helper out of the internationalisation section.
