+++ assets/pages/oneoff-contributions/oneoffContributions.jsx
@@ -18,7 +18,7 @@ import ContribLegal from 'components/legal/contribLegal/contribLegal';
 
 import { detect as detectCountry } from 'helpers/internationalisation/country';
 import { detect as detectCurrency } from 'helpers/internationalisation/currency';
-import { termsLinks } from 'helpers/internationalisation/legal';
+import { termsLinks } from 'helpers/legal';
 import * as user from 'helpers/user/user';
 import { getQueryParameter } from 'helpers/url';
 import { parse as parseContrib } from 'helpers/contributions';
